move error handling to service layer

+++ src/main/java/com/example/spring/jpa/services/MessageServiceImpl.java
@@ -2,7 +2,9 @@ package com.example.spring.jpa.services;
 
 import java.util.List;
 
+import com.example.spring.jpa.exceptions.UserMessageException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -26,21 +28,29 @@ public class MessageServiceImpl implements MessageService{
 
 	@Override
 	public UserMessage findById(long id) {
-		return userMessageRepository.findOne(id);
+		UserMessage message =  userMessageRepository.findOne(id);
+		if (message == null) throw new UserMessageException("message with id " + id + " not found", HttpStatus.NOT_FOUND);
+		return message;
 	}
 
 	@Override
 	public void saveMessage(UserMessage userMessage) {
+		if (isPresent(userMessage.getId())) throw new UserMessageException("Message with id " + userMessage.getId() + " already exist ", HttpStatus.BAD_REQUEST);
 		userMessageRepository.save(userMessage);
 	}
 
 	@Override
-	public void updateMessage(UserMessage userMessage){
-		saveMessage(userMessage);
+	public void updateMessage(long id, UserMessage userMessage){
+		userMessage.setId(id);
+		UserMessage currentMessage = findById(id);
+		if (currentMessage == null) throw new UserMessageException("Unable to update message with id " + id + " not found.", HttpStatus.NOT_FOUND);
+		currentMessage.setMessage(userMessage.getMessage());
+		userMessageRepository.save(currentMessage);
 	}
 
 	@Override
 	public void deleteMessageById(long id){
+		if (!isPresent(id)) throw new UserMessageException("Unable to delete message with id " + id + " not found.", HttpStatus.NOT_FOUND);
 		userMessageRepository.delete(id);
 	}
